Separating out namespace surfacemesh_detail for messy functions
